Added new station identifiers to gt.json and en.json.
The station identifiers "ASA", "FVW", "VDN", and "SPA" were added to both gt.json and en.json files to update the program with new stations. This is in line with the ongoing efforts to continually develop and maintain an updated database of locations for the racecs server backend. The names corresponding to the identifiers were also added accordingly.