Remove unused code and disable flight on player setup
Removed commented-out code related to targets in 'RaceManagement.java', as it's not used anymore, improving readability. Additionally, a player's ability to fly is now disabled during their setup in 'RaceSession.java'. This change was done to ensure the fairness of the game – no player should have the advantage of flying during racing sessions.