Refactor flight setting logic in RaceSession
Updated the RaceSession to set flight permissions for the entire team after finishing a race rather than the individual player. This changes the timing when flight is allowed. Also added checks in event handling methods to ensure operations are only performed on active and participant players. Exceptions were updated to make error messages more informative.

Modified division logic in RaceSessionTeams to allow teams of size 5 and explicit support for 10 players. This allows more flexibility in the number of players that can be divided equally into teams.